refactor: Move authentication middleware to a separate module for better organization
